Add tab header context menu implementation
